Rework TagSetModel and co.
We cannot use tag-sets in views that have vertices and edges that
do no implement the WithFeatures interface. So we have to instead
rely on some sort of tag-set projections, like for features.

With this the views properly receive a color generator that can
interpret the TAG settings for vertex and edge coloring (TODO,
deal properly with branch tags).

Right now there is not way to set or unset a tag: TODO.