Enfore the `ModelGraph` to be a simple, directed graph.
Since the edge pool in the model graph extends
`AbstractListenableEdgePool`, it is non-simple. We do not
want that first because this is not our target application
(following cells over time), second because we want to build
a branch graph on this model graph, and that the branch-graph
is only defined for simple graphs.